requirements.txt
psycopg2-binary==2.9.5
Django==4.2
djangorestframework==3.14.0
pillow==11.2.1
djangorestframework_simplejwt==5.5.0
django-grappelli==4.0.2